+++ src/integrations/text/plugin.rs
@@ -3,7 +3,9 @@ use bevy::{
     render::{Render, RenderApp, RenderSet, render_asset::RenderAssetPlugin},
 };
 
-use super::{VelloFont, font_loader::VelloFontLoader, render};
+use super::{
+    VelloFont, font_loader::VelloFontLoader, render, vello_text::calculate_text_content_size,
+};
 use crate::render::extract::VelloExtractStep;
 
 pub struct VelloTextIntegrationPlugin;
@@ -12,7 +14,8 @@ impl Plugin for VelloTextIntegrationPlugin {
     fn build(&self, app: &mut App) {
         app.init_asset::<VelloFont>()
             .init_asset_loader::<VelloFontLoader>()
-            .add_plugins(RenderAssetPlugin::<VelloFont>::default());
+            .add_plugins(RenderAssetPlugin::<VelloFont>::default())
+            .add_systems(Update, calculate_text_content_size);
 
         #[cfg(feature = "default_font")]
         {