+++ src/render/extract.rs
@@ -7,7 +7,7 @@ use bevy::{
 };
 
 use super::{VelloEntityCountData, VelloFrameProfileData};
-use crate::prelude::*;
+use crate::{VelloScreenSpace, prelude::*};
 
 #[derive(SystemSet, Debug, Hash, PartialEq, Eq, Clone)]
 pub enum VelloExtractStep {
@@ -22,6 +22,7 @@ pub struct ExtractedVelloScene {
     pub scene: VelloScene,
     pub transform: GlobalTransform,
     pub ui_node: Option<ComputedNode>,
+    pub screen_space: Option<VelloScreenSpace>,
 }
 
 pub fn extract_scenes(
@@ -39,6 +40,7 @@ pub fn extract_scenes(
                 &InheritedVisibility,
                 Option<&ComputedNode>,
                 Option<&RenderLayers>,
+                Option<&VelloScreenSpace>,
             ),
             Without<SkipEncoding>,
         >,
@@ -51,8 +53,15 @@ pub fn extract_scenes(
     let mut views: Vec<_> = query_views.iter().collect();
     views.sort_unstable_by_key(|(camera, _)| camera.order);
 
-    for (scene, transform, view_visibility, inherited_visibility, ui_node, render_layers) in
-        query_scenes.iter()
+    for (
+        scene,
+        transform,
+        view_visibility,
+        inherited_visibility,
+        ui_node,
+        render_layers,
+        screen_space,
+    ) in query_scenes.iter()
     {
         // Skip if visibility conditions are not met
         if !view_visibility.get() || !inherited_visibility.get() {
@@ -69,6 +78,7 @@ pub fn extract_scenes(
                     transform: *transform,
                     scene: scene.clone(),
                     ui_node: ui_node.cloned(),
+                    screen_space: screen_space.cloned(),
                 })
                 .insert(TemporaryRenderEntity);
             n_scenes += 1;